+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/dtos/response/UserDTO.java
@@ -1,14 +1,13 @@
 package com.swp391.school_medical_management.modules.users.dtos.response;
 
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.swp391.school_medical_management.modules.users.entities.UserEntity.UserRole;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 
 @Data
@@ -16,8 +15,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class UserDTO{
-    private Long id;
+public class UserDTO {
+    private int id;
     private String fullName;
     private String email;
     private String password;
