+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/repositories/VaccineNameRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.swp391.school_medical_management.modules.users.entities.VaccineNameEntity;
 
-public interface VaccineNameRepository extends JpaRepository<VaccineNameEntity, Long> {
+public interface VaccineNameRepository extends JpaRepository<VaccineNameEntity, Integer> {
     // Optional<VaccineNameEntity> findByVaccineName(String vaccineName); 
-    // Optional<VaccineNameEntity> findById(Long id);                     
-    // Optional<VaccineNameEntity> findByVaccineNameAndManufactureAndUrlAndNote (String name, String url,String manufacture, String note);         
+    // Optional<VaccineNameEntity> findByVaccineNameAndManufactureAndUrlAndNote (String name, String url,String manufacture, String note);
 }
